update file structure based on notes from professor
