+++ Assignment2/View/CryptoItem.swift
