+++ Assignment2/View/CurrencyItem.swift
