+++ Assignment2/View/NavigationMenu.swift
